Be compatible with Python 2.6 AND 2.7
Remember when I "fixed" this script on Python 2.6 / OS X 10.6? That
broke it on Python 2.7 / OS X 10.7.

Whose irresponsible choice was it to change the array values in
platform.python_version_tuple() from integers to strings in a dot
release?

My previous "update" didn't get a new revision number, so this one got
bumped two to be technically correct.